Fix LXC detect in boostrap-fixup.in for newer Linux host systems
Newer host systems such as Debian 12 use control group v2 and so the heuristics for this `if` statement that attempts to detect LXC fail on such systems. 

This additional `||` clause checks the process name of PID 1 which is `init` outside a container and `lxc-execute` inside a container.

This is tested and works on Debian-12 and also does not introduce any regressions.